+++ app/javascript/components/widget/widget-templates/area-coverage/component.js
@@ -1,0 +1,28 @@
+import React, { Fragment } from 'react';
+import PropTypes from 'prop-types';
+
+import Chart from 'components/recharts';
+
+import config from './config';
+
+const AreaCoverage = ({ widgetData }) => {
+  if (!widgetData) return null;
+  const data = config.parse(widgetData);
+  const { chartConfig, chartData } = data;
+
+  return (
+    <Fragment>
+      <div className="c-widget-template">
+
+        <Chart
+          data={chartData}
+          config={chartConfig}
+        />
+      </div>
+    </Fragment>
+  );
+};
+
+AreaCoverage.propTypes = { widgetData: PropTypes.array.isRequired };
+
+export default AreaCoverage;